+++ R/airsis_loadDaily.R
@@ -1,4 +1,5 @@
 #' @export
+#' @importFrom dplyr across
 #'
 #' @title Load daily AIRSIS monitoring data
 #'
@@ -122,9 +123,11 @@ airsis_loadDaily <- function(
 
   data <-
     data %>%
-    dplyr::select(dplyr::all_of(c("datetime", meta$deviceDeploymentID)))
+    dplyr::select(dplyr::all_of(c("datetime", meta$deviceDeploymentID))) %>%
+    # Replace any NaN that snuck in
+    dplyr::mutate(across(where(is.numeric), function(x) ifelse(is.nan(x), NA, x)))
 
-  # Create monitor objecet
+  # Create monitor object
   monitor <- list(meta = meta, data = data)
 
   monitor <- structure(monitor, class = c("mts_monitor", "mts", class(monitor)))