+++ R/data.R
@@ -12,7 +12,7 @@
 #' northern Idaho. The NW_Megafires dataset contains monitoring data for the
 #' Pacific Northwest from July 24 through September 06, 2015.
 #'
-#' This dataset was generated on 2022-03-22 by running:
+#' This dataset was generated on 2022-03-23 by running:
 #'
 #' \preformatted{
 #' library(AirMonitor)